+++ src/pages/basic-data/nvi/publication-points/NviAdminPublicationPointsPage.tsx
@@ -6,7 +6,7 @@ import {
 } from '../../../../components/dialogs/helper-texts/NviPointsHelperTextModal';
 import { NviPageLayout } from '../../../../components/page-layouts/NviPageLayout';
 import { TableSkeleton } from '../../../../components/skeletons/TableSkeleton';
-import { NviAdminSortSelectorType } from '../../../../components/sort-selectors/sort-nvi-table/nvi-admin-sort-types';
+import { NviAdminSortSelectorType } from '../components/utils/nvi-admin-sort-types';
 import { HorizontalBox, VerticalBox } from '../../../../components/styled/Wrappers';
 import { CenteredTableCell } from '../../../../components/tables/table-styles';
 import { InstitutionReport } from '../../../../types/nvi.types';
